Rename the Marketplace Product Installation folder
