+++ client/my-sites/marketplace/controller.js
@@ -2,7 +2,7 @@ import page from 'page';
 import { getSiteFragment } from 'calypso/lib/route';
 import MarketplaceThankYou from 'calypso/my-sites/checkout/checkout-thank-you/marketplace/marketplace-thank-you';
 import MarketplaceDomainUpsell from 'calypso/my-sites/marketplace/pages/marketplace-domain-upsell';
-import MarketplacePluginInstall from 'calypso/my-sites/marketplace/pages/marketplace-plugin-install';
+import MarketplaceProductInstall from 'calypso/my-sites/marketplace/pages/marketplace-product-install';
 import MarketplaceTest from 'calypso/my-sites/marketplace/pages/marketplace-test';
 import SignupSuccess from 'calypso/my-sites/marketplace/pages/submission-success/signup-success';
 
@@ -13,13 +13,13 @@ export function renderDomainsPage( context, next ) {
 
 export function renderPluginsInstallPage( context, next ) {
 	const { productSlug } = context.params;
-	context.primary = <MarketplacePluginInstall productSlug={ productSlug } />;
+	context.primary = <MarketplaceProductInstall productSlug={ productSlug } />;
 	next();
 }
 
 export function renderThemesInstallPage( context, next ) {
 	const { themeSlug } = context.params;
-	context.primary = <MarketplacePluginInstall themeSlug={ themeSlug } />;
+	context.primary = <MarketplaceProductInstall themeSlug={ themeSlug } />;
 	next();
 }
